Refactored DB Examples to the Debug endpoint
